style(nav-brick): update nav menu Refs: NEXT_BUILDER-1330
