+++ bricks/frame-bricks/src/index.ts
@@ -1,3 +1,4 @@
 import "./i18n";
 import "./side-bar";
 import "./nav-bar";
+import "./nav-menu";
