+++ bricks/frame-bricks/src/nav-bar/NavBar.shadow.css
@@ -1,0 +1,27 @@
+.nav-bar {
+  position: fixed;
+  top: 0;
+  right: 0;
+  left: 0;
+  height: 32px;
+  padding: 10px 16px;
+  display: flex;
+  background-color: var(--antd-component-background);
+}
+
+.left-content,
+.right-content {
+  display: flex;
+  flex: 1;
+  line-height: 32px;
+}
+
+.nav-menu,
+.nav-buttons {
+  flex: 1;
+}
+
+.nav-buttons {
+  display: flex;
+  justify-content: flex-end;
+}
