+++ bricks/frame-bricks/src/nav-menu/index.spec.ts
@@ -1,0 +1,20 @@
+import ReactDOM from "react-dom";
+import "./";
+
+const spyOnRender = jest
+  .spyOn(ReactDOM, "render")
+  .mockImplementation(() => null);
+const unmountComponentAtNode = jest
+  .spyOn(ReactDOM, "unmountComponentAtNode")
+  .mockImplementation(() => null);
+
+describe("frame-bricks.nav-menu", () => {
+  it("should create a custom element", () => {
+    const element = document.createElement("frame-bricks.nav-menu");
+    expect(spyOnRender).not.toBeCalled();
+    document.body.appendChild(element);
+    expect(spyOnRender).toBeCalled();
+    document.body.removeChild(element);
+    expect(unmountComponentAtNode).toBeCalled();
+  });
+});
